♻️ Refactor imports in blend and core modules; update load_equilibrium function to raise RuntimeError for missing rho_tor_norm
